# hfitd/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: daniele@smartango.com
    name: Daniele Cruciani
  description: API for managing database fixtures, file operations, and package templates
  license:
    name: MIT
    url: https://github.com/danielecr/hot-fixture-tool/blob/main/LICENSE
  termsOfService: https://github.com/danielecr/hot-fixture-tool
  title: Hot Fixture Tool Daemon API
  version: 1.0.0
paths:
...
  /db/{dbms}/{dbid}/table/{tableid}/rows:
    get:
      consumes:
      - application/json
      description: Get rows from a specific table with optional filtering
      parameters:
      - description: DBMS provider name
        example: mysql
        in: path
        name: dbms
        required: true
        type: string
      - description: Database name
        example: myapp_db
        in: path
        name: dbid
        required: true
        type: string
      - description: Table name
        example: users
        in: path
        name: tableid
        required: true
        type: string
      - description: SQL filter clause
        example: '"WHERE id > 100 ORDER BY name DESC"'
        in: query
        name: filterpart
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Table rows
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get table rows
      tags:
      - database
  /db/{dbms}/{dbid}/tables:
    get:
      consumes:
      - application/json
      description: Get list of all tables in a specific database
      parameters:
      - description: DBMS provider name
        example: mysql
        in: path
        name: dbms
        required: true
        type: string
      - description: Database name
        example: myapp_db
        in: path
        name: dbid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of table names
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad request - invalid DBMS or database
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List tables
      tags:
      - database
  /db/{dbms}/dbs:
    get:
      consumes:
      - application/json
      description: Get list of all databases for a specific DBMS
      parameters:
      - description: DBMS provider name
        example: mysql
        in: path
        name: dbms
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of database names
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad request - invalid DBMS
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List databases
      tags:
      - database
  /db/dbmss:
    get:
      consumes:
      - application/json
      description: Get list of all available database management systems
      produces:
      - application/json
      responses:
        "200":
          description: List of DBMS providers
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List DBMS providers
      tags:
      - database
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"